Revert "Format change"
