--- glog_file_linux.go
@@ -1,39 +1,0 @@
-// Go support for leveled logs, analogous to https://github.com/google/glog.
-//
-// Copyright 2023 Google Inc. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-//go:build linux
-
-package glog
-
-import (
-	"errors"
-	"runtime"
-	"syscall"
-)
-
-// abortProcess attempts to kill the current process in a way that will dump the
-// currently-running goroutines someplace useful (like stderr).
-//
-// It does this by sending SIGABRT to the current thread.
-//
-// If successful, abortProcess does not return.
-func abortProcess() error {
-	runtime.LockOSThread()
-	if err := syscall.Tgkill(syscall.Getpid(), syscall.Gettid(), syscall.SIGABRT); err != nil {
-		return err
-	}
-	return errors.New("log: killed current thread with SIGABRT, but still running")
-}
