+++ glog_flags.go
@@ -133,11 +133,6 @@ func (l *Level) Set(value string) error {
 type vModuleFlag struct{ *verboseFlags }
 
 func (f vModuleFlag) String() string {
-	// Do not panic on the zero value.
-	// https://groups.google.com/g/golang-nuts/c/Atlr8uAjn6U/m/iId17Td5BQAJ.
-	if f.verboseFlags == nil {
-		return ""
-	}
 	f.mu.Lock()
 	defer f.mu.Unlock()
 
@@ -197,7 +192,9 @@ func (f *verboseFlags) levelForPC(pc uintptr) Level {
 	file, _ := fn.FileLine(pc)
 	// The file is something like /a/b/c/d.go. We want just the d for
 	// regular matches, /a/b/c/d for full matches.
-	file = strings.TrimSuffix(file, ".go")
+	if strings.HasSuffix(file, ".go") {
+		file = file[:len(file)-3]
+	}
 	full := file
 	if slash := strings.LastIndex(file, "/"); slash >= 0 {
 		file = file[slash+1:]
