+++ go.mod
@@ -1,3 +1,3 @@
 module github.com/golang/glog
 
-go 1.19
+go 1.18
